- [x] Refactor: Break up `theme-quadrone.svelte.ts` - the style generation code could live in its file, etc.
Archived Theme Setting TODOs and notes. Made plans.

Fixed errors during testing.
